+++ natuur/src/components/Authorization/Information/InformationDisableInputBox.tsx
@@ -1,22 +1,18 @@
 import React, { FC } from "react";
 
-import AcceptDisableButton from "./AcceptDisableButton";
+import AcceptDisableButton from "../Accept/AcceptDisableButton";
 import {
   InfomationInputBoxCover,
   GradationHorizon
-} from "../../styles/Authorization";
-import InputRow from "../default/Common/InputRow";
+} from "../../../styles/Authorization";
+import InputRow from "../../default/Common/InputRow";
 
 interface Props {
   updateToastr(toastrInformation: object): void;
-  removeToastr(): void;
 }
 
-const InformationDisableInputBox: FC<Props> = ({
-  updateToastr,
-  removeToastr
-}) => (
-  <InfomationInputBoxCover isDisable={true}>
+const InformationDisableInputBox: FC<Props> = ({ updateToastr }) => (
+  <InfomationInputBoxCover isDisable>
     <div>
       {/* 이메일 입력란 */}
       <InputRow
@@ -24,11 +20,10 @@ const InformationDisableInputBox: FC<Props> = ({
         placeHolder="example@dsmhs.kr"
         name="emailInput"
         type="text"
-        isDisable={true}
-        isReadOnly={true}
+        isReadOnly
       />
 
-      <GradationHorizon isDisable={true} />
+      <GradationHorizon />
 
       {/* 비밀번호 입력란 */}
       <InputRow
@@ -36,13 +31,12 @@ const InformationDisableInputBox: FC<Props> = ({
         placeHolder="••••••••"
         name="passwordInput"
         type="password"
-        isDisable={true}
-        isWarning={true}
-        isReadOnly={true}
+        isWarning
+        isReadOnly
         warningMessage="* 영문, 숫자 포함 8자리 이상 16자리 이하"
       />
 
-      <GradationHorizon isDisable={true} />
+      <GradationHorizon />
 
       {/* 비밀번호 (확인) 입력란 */}
       <InputRow
@@ -50,14 +44,10 @@ const InformationDisableInputBox: FC<Props> = ({
         placeHolder="••••••••"
         name="passwordCheckInput"
         type="password"
-        isDisable={true}
-        isReadOnly={true}
+        isReadOnly
       />
     </div>
-    <AcceptDisableButton
-      updateToastr={updateToastr}
-      removeToastr={removeToastr}
-    />
+    <AcceptDisableButton updateToastr={updateToastr} />
   </InfomationInputBoxCover>
 );
